Bug 1498185 - Move versioning check to spec validation within registry package
